collision bug fix

+++ Assets/Scripts/InGame/BattleFields/Bullets/BulletManager.cs
@@ -40,22 +40,23 @@ namespace InGame.BattleFields.Bullets
         private BulletSetUp AddNumericalBuffs(BulletSetUp bulletSetUp, WeaponBuff buff)
         {
             // life time
-            bulletSetUp.lifeTime *= 1 + (float)buff.lifeTimePercentageBuff/100f;
+            bulletSetUp.lifeTime *= 1 + buff.lifeTimePercentageBuff / 100f;
             bulletSetUp.lifeTime = Mathf.Max(0.001f, bulletSetUp.lifeTime);
 
             // speed
-            bulletSetUp.speed *= 1 + (float)buff.speedPercentageBuff/100f;
+            bulletSetUp.speed *= 1 + buff.speedPercentageBuff / 100f;
             bulletSetUp.speed = Mathf.Max(0.001f, bulletSetUp.speed);
 
             // damage
             // TODO bulletSetUp.damage *= m_damageMultiplier.value;
-            bulletSetUp.damage *= 1 + (float)buff.damagePercentageBuff/100f;
+            bulletSetUp.damage *= 1 + buff.damagePercentageBuff / 100f;
             bulletSetUp.damage += buff.flatDamageBuff;
             bulletSetUp.damage = Mathf.Max(0.001f, bulletSetUp.damage);
 
             // size
             bulletSetUp.size *= 1 + (float)buff.bulletSizePercentageBuff/100f;
             bulletSetUp.size = Mathf.Max(0.001f, bulletSetUp.size);
+
             return bulletSetUp;
         }
